requirements.txt
streamlit
google-generativeai
gTTS
pydub